math1.py
# math_tool_server.py
from mcp.server.fastmcp import FastMCP

mcp = FastMCP("MathTools")

@mcp.tool()
def add(a: int, b: int) -> int:
    """Add two numbers"""
    return a + b

if __name__ == "__main__":
    mcp.run(transport="streamable-http")


