requirements.txt
langchain
langgraph
langchain-google-genai
langchain-mcp-adapters
python-dotenv
